add gradient functionality
Added function plot_gradient creates plot where field values are projected onto a given line. The line is defined by start and end point. The line is separated into subpoints (number can be chosen by user) and the field value, which is next to the suboints is used. An interpolation, where the distance of the nodes to the subpoints is used to calculate a weighted value didn't worked out so well, therefore this simple approach.
A poly line can be fitted to the resulting gradient and the equation is shown in the plot legend.

Added function get_relative_gradient calculates the relative (e.g. stress) gradient via a fitted poly line. Parts of the above function are used.